# backend/cmd/server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for managing financial data, chat interactions, and scenario planning",
        "title": "Financial Chat System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/v2/assets": {
            "get": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Returns paginated non-cash assets for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets V2"
                ],
                "summary": "List assets (v2)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max items to return (-1 for all)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Creates a new non-cash asset for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets V2"
                ],
                "summary": "Create an asset (v2)",
                "parameters": [
                    {
                        "description": "Asset data",
                        "name": "asset",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.assetCreateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repository.NonCashAsset"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Bulk deletes all non-cash assets for the authenticated user in a single query.",
                "tags": [
                    "Bulk Delete V2"
                ],
                "summary": "Delete all assets (v2)",
                "operationId": "deleteBulkAssets",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/v2/cash-accounts": {
            "get": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Returns paginated cash accounts for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cash Accounts V2"
                ],
                "summary": "List cash accounts (v2)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max items to return (-1 for all)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Bulk deletes all cash accounts for the authenticated user in a single query.",
                "tags": [
                    "Bulk Delete V2"
                ],
                "summary": "Delete all cash accounts (v2)",
                "operationId": "deleteBulkCashAccounts",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/v2/cashflow/incomes": {
            "get": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Returns paginated incomes for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Incomes V2"
                ],
                "summary": "List incomes (v2)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max items to return (-1 for all)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Creates a new income for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Incomes V2"
                ],
                "summary": "Create an income (v2)",
                "parameters": [
                    {
                        "description": "Income data",
                        "name": "income",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.incomeV2CreateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financial-chat-system_backend_internal_financial_v2_repository.Income"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Bulk deletes all incomes for the authenticated user in a single query.",
                "tags": [
                    "Bulk Delete V2"
                ],
                "summary": "Delete all incomes (v2)",
                "operationId": "deleteBulkIncomes",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/v2/cpf/accounts": {
            "get": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Returns all CPF accounts for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CPF V2"
                ],
                "summary": "List CPF accounts (v2)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repository.CPFAccount"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Bulk deletes all CPF accounts for the authenticated user in a single query.",
                "tags": [
                    "Bulk Delete V2"
                ],
                "summary": "Delete all CPF accounts (v2)",
                "operationId": "deleteBulkCPFAccounts",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/v2/investments": {
            "get": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Returns paginated investments for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Investments V2"
                ],
                "summary": "List investments (v2)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max items to return (-1 for all)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Creates a new investment for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Investments V2"
                ],
                "summary": "Create an investment (v2)",
                "parameters": [
                    {
                        "description": "Investment data",
                        "name": "investment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.investmentCreateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financial-chat-system_backend_internal_financial_v2_repository.Investment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Bulk deletes all investments for the authenticated user in a single query.",
                "tags": [
                    "Bulk Delete V2"
                ],
                "summary": "Delete all investments (v2)",
                "operationId": "deleteBulkInvestments",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/v2/liabilities": {
            "get": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Returns paginated liabilities for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Liabilities V2"
                ],
                "summary": "List liabilities (v2)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max items to return (-1 for all)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Creates a new liability. If minimumPayment \u003e 0, automatically creates a linked expense for debt repayment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Liabilities V2"
                ],
                "summary": "Create a liability (v2)",
                "parameters": [
                    {
                        "description": "Liability to create",
                        "name": "liability",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.liabilityCreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/financial-chat-system_backend_internal_financial_v2_repository.Liability"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Bulk deletes all liabilities for the authenticated user in a single query.",
                "tags": [
                    "Bulk Delete V2"
                ],
                "summary": "Delete all liabilities (v2)",
                "operationId": "deleteBulkLiabilities",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/v2/reset-all-data": {
            "delete": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Deletes all financial data (assets, liabilities, incomes, expenses, investments,",
                "tags": [
                    "Bulk Delete V2"
                ],
                "summary": "Reset all user data (v2)",
                "operationId": "resetAllUserData",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.assetCreateInput": {
            "type": "object",
            "properties": {
                "annualGrowthRate": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "currentValue": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "growthStrategy": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "terminalValue": {
                    "type": "string"
                }
            }
        },
...
        "handlers.incomeV2CreateInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "cpfWageType": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "frequency": {
                    "type": "string"
                },
                "growthRate": {
                    "type": "string"
                },
                "growthStrategy": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "personId": {
                    "description": "Required FK to persons table",
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AuthToken": {
            "description": "JWT token for authentication (HMAC-signed with BACKEND_SHARED_SECRET). In dev mode, use GET /api/dev/token?user_id=\u003cuuid\u003e to generate one.",
            "type": "apiKey",
            "name": "X-Auth-Token",
            "in": "header"
        }
    }
}